requirements.txt
pytest==2.7.2
pytest-describe==0.10.3
pyhamcrest==1.9.0
mock==1.3.0
coverage==4.0.3
flake8==3.0.4
